+++ fmc-master/FMC/software/Car/encode.h
@@ -4,8 +4,8 @@
 #include "headfile.h"
 #include "struct.h"
 
-extern int8_t getCountNum_L;  //获取编码器计数值
-extern int8_t getCountNum_R;  //获取编码器计数值
+extern int16_t getCountNum_L;  //获取编码器计数值
+extern int16_t getCountNum_R;  //获取编码器计数值
 
 
 void encode_init(void);
